Persist in Redis to survive crashes
I created a `cast` for getting the initial store from redis so as not to slow down the `init` callback, which is synchronous. As long as we don’t start the heartbeat until after initializing the store, we should be good.

You can test in the console by closing and opening another chat console: you’ll still be reminded.